+++ src/core/system/dynamic_library.cpp
@@ -5,9 +5,11 @@
 #    include <winnt.h>
 #endif
 #ifdef NEFORCE_PLATFORM_LINUX
+#    include <linux/limits.h>
 #    include <dlfcn.h>
 #    include <elf.h>
 #    include <link.h>
+#    include <unistd.h>
 #endif
 NEFORCE_BEGIN_NAMESPACE__
 
@@ -167,7 +169,7 @@ dynamic_library dynamic_library::load_by_name(const string& name, load_mode mode
         if (!name.starts_with("lib")) {
             resolved_name = "lib" + name;
         }
-        if (name.size() < 3 || !name.ends_with(".so")) {
+        if (name.size() < 3 || (!name.ends_with(".so") && name.find(".so.") == string::npos)) {
             resolved_name += ".so";
         }
 #endif
@@ -182,7 +184,7 @@ vector<string> dynamic_library::list_symbols(const string& name_filter) const {
     }
 
 #ifdef NEFORCE_PLATFORM_LINUX
-    struct ::link_map* lm = nullptr;
+    ::link_map* lm = nullptr;
     if (::dlinfo(handle_, RTLD_DI_LINKMAP, &lm) != 0 || lm == nullptr) {
         return symbols;
     }
@@ -245,9 +247,9 @@ vector<string> dynamic_library::list_symbols(const string& name_filter) const {
     if (strtab != nullptr && symtab != nullptr && syment > 0 && nchain > 0) {
         for (size_t i = 1; i < nchain; ++i) {
             const auto* sym = reinterpret_cast<const ElfW(Sym)*>(reinterpret_cast<const char*>(symtab) + i * syment);
-            const char* sym_name = strtab + sym->st_name;
+            string_view sym_name = strtab + sym->st_name;
             if (sym_name[0] != '\0' && ELF32_ST_TYPE(sym->st_info) == STT_FUNC) {
-                if (name_filter.empty() || string_view(sym_name).find(name_filter) != string::npos) {
+                if (name_filter.empty() || sym_name.contains(name_filter.view())) {
                     symbols.push_back(sym_name);
                 }
             }